requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# MOPS Sheets Uploader Dependencies
# Core dependencies for the MOPS Sheets Uploader package

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# Google Sheets integration
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# Configuration and file handling
PyYAML>=6.0
python-dotenv>=0.19.0

# Date and time handling
python-dateutil>=2.8.0

# Logging and utilities
colorlog>=6.6.0

# Optional: Enhanced CSV handling
chardet>=4.0.0

# Development and testing dependencies (optional)
# Uncomment for development environment
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991

# Documentation dependencies (optional)
# Uncomment for documentation generation  
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0